Deduplicate sql driver helpers and http internals; trim url/fetch glue
