+++ src/js/internal/http/FakeSocket.ts
@@ -1,4 +1,4 @@
-const { kInternalSocketData, serverSymbol } = require("internal/http");
+const { kInternalSocketData, serverSymbol, installSocketStubs } = require("internal/http");
 const { kAutoDestroyed } = require("internal/shared");
 const { Duplex } = require("internal/stream");
 
@@ -24,13 +24,6 @@ var FakeSocket = class Socket extends Duplex {
       (internalData = this[kInternalSocketData])?.[0]?.[serverSymbol]?.requestIP(internalData[2]) ?? {});
   }
 
-  get bufferSize() {
-    return this.writableLength;
-  }
-
-  connect(_port, _host, _connectListener) {
-    return this;
-  }
   _onTimeout = function () {
     this.emit("timeout");
   };
@@ -55,59 +48,7 @@ var FakeSocket = class Socket extends Duplex {
     return 80;
   }
 
-  get pending() {
-    return this.connecting;
-  }
-
   _read(_size) {}
-
-  get readyState() {
-    if (this.connecting) return "opening";
-    if (this.readable) {
-      return this.writable ? "open" : "readOnly";
-    } else {
-      return this.writable ? "writeOnly" : "closed";
-    }
-  }
-
-  ref() {
-    return this;
-  }
-
-  get remoteAddress() {
-    return this.address()?.address;
-  }
-
-  set remoteAddress(val) {
-    // initialize the object so that other properties wouldn't be lost
-    this.address().address = val;
-  }
-
-  get remotePort() {
-    return this.address()?.port;
-  }
-
-  set remotePort(val) {
-    // initialize the object so that other properties wouldn't be lost
-    this.address().port = val;
-  }
-
-  get remoteFamily() {
-    return this.address()?.family;
-  }
-
-  set remoteFamily(val) {
-    // initialize the object so that other properties wouldn't be lost
-    this.address().family = val;
-  }
-
-  resetAndDestroy() {}
-
-  setKeepAlive(_enable = false, _initialDelay = 0) {}
-
-  setNoDelay(_noDelay = true) {
-    return this;
-  }
 
   setTimeout(timeout, callback) {
     const socketData = this[kInternalSocketData];
@@ -115,10 +56,6 @@ var FakeSocket = class Socket extends Duplex {
 
     const http_res = socketData[1];
     http_res?.req?.setTimeout(timeout, callback);
-    return this;
-  }
-
-  unref() {
     return this;
   }
 
@@ -130,6 +67,7 @@ var FakeSocket = class Socket extends Duplex {
   }
 };
 
+installSocketStubs(FakeSocket);
 Object.defineProperty(FakeSocket, "name", { value: "Socket" });
 
 export default {
