+++ src/js/internal/sql/postgres.ts
@@ -1,17 +1,18 @@
 import type { PostgresErrorOptions } from "internal/sql/errors";
 import type { Query } from "./query";
-import type { ArrayType, DatabaseAdapter, SQLArrayParameter, SQLHelper, SQLResultArray, SSLMode } from "./shared";
+import type { ArrayType, DatabaseAdapter, SQLArrayParameter, SQLCommand, SQLResultArray, SSLMode } from "./shared";
 const {
-  SQLHelper,
   SSLMode,
   SQLResultArray,
   SQLArrayParameter,
-  buildDefinedColumnsAndQuery,
+  BasePooledConnection,
+  BaseSQLAdapter,
+  getHelperCommandFromDetect,
+  pushBindParam,
 } = require("internal/sql/shared");
 const {
-  Query,
   SQLQueryFlags,
-  symbols: { _strings, _values, _flags, _results, _handle },
+  symbols: { _results, _handle },
 } = require("internal/sql/query");
 function isTypedArray(value: any) {
   // Buffer should be treated as a normal object
@@ -346,122 +347,7 @@ export interface PostgresDotZig {
   ) => $ZigGeneratedClasses.PostgresSQLQuery;
 }
 
-const enum SQLCommand {
-  insert = 0,
-  update = 1,
-  updateSet = 2,
-  where = 3,
-  in = 4,
-  none = -1,
-}
-export type { SQLCommand };
-
-function commandToString(command: SQLCommand): string {
-  switch (command) {
-    case SQLCommand.insert:
-      return "INSERT";
-    case SQLCommand.updateSet:
-    case SQLCommand.update:
-      return "UPDATE";
-    case SQLCommand.in:
-    case SQLCommand.where:
-      return "WHERE";
-    default:
-      return "";
-  }
-}
-
-function detectCommand(query: string): SQLCommand {
-  const text = query.toLowerCase().trim();
-  const text_len = text.length;
-
-  let token = "";
-  let command = SQLCommand.none;
-  let quoted = false;
-  // we need to reverse search so we find the closest command to the parameter
-  for (let i = text_len - 1; i >= 0; i--) {
-    const char = text[i];
-    switch (char) {
-      case " ": // Space
-      case "\n": // Line feed
-      case "\t": // Tab character
-      case "\r": // Carriage return
-      case "\f": // Form feed
-      case "\v": {
-        switch (token) {
-          case "insert": {
-            return SQLCommand.insert;
-          }
-          case "update": {
-            return SQLCommand.update;
-          }
-          case "where": {
-            return SQLCommand.where;
-          }
-          case "set": {
-            return SQLCommand.updateSet;
-          }
-          case "in": {
-            return SQLCommand.in;
-          }
-          default: {
-            token = "";
-            continue;
-          }
-        }
-      }
-      default: {
-        // skip quoted commands
-        if (char === '"') {
-          quoted = !quoted;
-          continue;
-        }
-        if (!quoted) {
-          token = char + token;
-        }
-      }
-    }
-  }
-  if (token) {
-    switch (token) {
-      case "insert":
-        return SQLCommand.insert;
-      case "update":
-        return SQLCommand.update;
-      case "where":
-        return SQLCommand.where;
-      case "set":
-        return SQLCommand.updateSet;
-      case "in":
-        return SQLCommand.in;
-      default:
-        return SQLCommand.none;
-    }
-  }
-  return command;
-}
-
-const enum PooledConnectionState {
-  pending = 0,
-  connected = 1,
-  closed = 2,
-}
-
-const enum PooledConnectionFlags {
-  /// canBeConnected is used to indicate that at least one time we were able to connect to the database
-  canBeConnected = 1 << 0,
-  /// reserved is used to indicate that the connection is currently reserved
-  reserved = 1 << 1,
-  /// preReserved is used to indicate that the connection will be reserved in the future when queryCount drops to 0
-  preReserved = 1 << 2,
-}
-
-function onQueryFinish(this: PooledPostgresConnection, onClose: (err: Error) => void) {
-  this.queries.delete(onClose);
-  this.adapter.release(this);
-}
-
-class PooledPostgresConnection {
+class PooledPostgresConnection extends BasePooledConnection<$ZigGeneratedClasses.PostgresSQLConnection> {
   private static async createConnection(
     options: Bun.SQL.__internal.DefinedPostgresOrMySQLOptions,
     onConnected: (err: Error | null, connection: $ZigGeneratedClasses.PostgresSQLConnection) => void,
@@ -520,156 +406,41 @@ class PooledPostgresConnection {
     }
   }
 
-  adapter: PostgresAdapter;
-  connection: $ZigGeneratedClasses.PostgresSQLConnection | null = null;
-  state: PooledConnectionState = PooledConnectionState.pending;
-  storedError: Error | null = null;
-  queries: Set<(err: Error) => void> = new Set();
-  onFinish: ((err: Error | null) => void) | null = null;
-  connectionInfo: Bun.SQL.__internal.DefinedPostgresOrMySQLOptions;
-  flags: number = 0;
-  /// queryCount is used to indicate the number of queries using the connection, if a connection is reserved or if its a transaction queryCount will be 1 independently of the number of queries
-  queryCount: number = 0;
-
-  #onConnected(err, _) {
-    if (err) {
-      err = wrapPostgresError(err);
-    }
-    const connectionInfo = this.connectionInfo;
-    if (connectionInfo?.onconnect) {
-      connectionInfo.onconnect(err);
-    }
-    this.storedError = err;
-    if (!err) {
-      this.flags |= PooledConnectionFlags.canBeConnected;
-    }
-    this.state = err ? PooledConnectionState.closed : PooledConnectionState.connected;
-    const onFinish = this.onFinish;
-    if (onFinish) {
-      this.queryCount = 0;
-      this.flags &= ~PooledConnectionFlags.reserved;
-      this.flags &= ~PooledConnectionFlags.preReserved;
-
-      // pool is closed, lets finish the connection
-      // pool is closed, lets finish the connection
-      if (err) {
-        onFinish(err);
-      } else {
-        this.connection?.close();
-      }
-      return;
-    }
-    this.adapter.release(this, true);
-  }
-
-  #onClose(err) {
-    if (err) {
-      err = wrapPostgresError(err);
-    }
-    const connectionInfo = this.connectionInfo;
-    if (connectionInfo?.onclose) {
-      connectionInfo.onclose(err);
-    }
-    this.state = PooledConnectionState.closed;
-    this.connection = null;
-    this.storedError = err;
-
-    // remove from ready connections if its there
-    this.adapter.readyConnections?.delete(this);
-    const queries = new Set(this.queries);
-    this.queries?.clear?.();
-    this.queryCount = 0;
-    this.flags &= ~PooledConnectionFlags.reserved;
-
-    // notify all queries that the connection is closed
-    for (const onClose of queries) {
-      onClose(err);
-    }
-    const onFinish = this.onFinish;
-    if (onFinish) {
-      onFinish(err);
-    }
-
-    this.adapter.release(this, true);
-  }
-
-  constructor(connectionInfo: Bun.SQL.__internal.DefinedPostgresOrMySQLOptions, adapter: PostgresAdapter) {
-    this.state = PooledConnectionState.pending;
-    this.adapter = adapter;
-    this.connectionInfo = connectionInfo;
-    this.#startConnection();
-  }
-
-  async #startConnection() {
+  protected async startConnection() {
     this.connection = await PooledPostgresConnection.createConnection(
       this.connectionInfo,
-      this.#onConnected.bind(this),
-      this.#onClose.bind(this),
+      this.handleConnected.bind(this),
+      this.handleClose.bind(this),
     );
   }
 
-  onClose(onClose: (err: Error) => void) {
-    this.queries.add(onClose);
+  protected wrapError(error: any): Error {
+    return wrapPostgresError(error);
   }
 
-  bindQuery(query: Query<any, any>, onClose: (err: Error) => void) {
-    this.queries.add(onClose);
-    query.finally(onQueryFinish.bind(this, onClose));
-  }
-
-  #doRetry() {
-    if (this.adapter.closed) {
-      return;
+  protected isNonRetryableError(code: string | undefined): boolean {
+    switch (code) {
+      case "ERR_POSTGRES_UNSUPPORTED_AUTHENTICATION_METHOD":
+      case "ERR_POSTGRES_UNKNOWN_AUTHENTICATION_METHOD":
+      case "ERR_POSTGRES_TLS_NOT_AVAILABLE":
+      case "ERR_POSTGRES_TLS_UPGRADE_FAILED":
+      case "ERR_POSTGRES_INVALID_SERVER_SIGNATURE":
+      case "ERR_POSTGRES_INVALID_SERVER_KEY":
+      case "ERR_POSTGRES_AUTHENTICATION_FAILED_PBKDF2":
+        // we can't retry these are authentication errors
+        return true;
+      default:
+        return false;
     }
-    // reset error and state
-    this.storedError = null;
-    this.state = PooledConnectionState.pending;
-    // retry connection
-    this.#startConnection();
-  }
-  close() {
-    try {
-      if (this.state === PooledConnectionState.connected) {
-        this.connection?.close();
-      }
-    } catch {}
-  }
-  flush() {
-    this.connection?.flush();
-  }
-  retry() {
-    // if pool is closed, we can't retry
-    if (this.adapter.closed) {
-      return false;
-    }
-    // we need to reconnect
-    // lets use a retry strategy
-
-    // we can only retry if one day we are able to connect
-    if (this.flags & PooledConnectionFlags.canBeConnected) {
-      this.#doRetry();
-    } else {
-      // analyse type of error to see if we can retry
-      switch (this.storedError?.code) {
-        case "ERR_POSTGRES_UNSUPPORTED_AUTHENTICATION_METHOD":
-        case "ERR_POSTGRES_UNKNOWN_AUTHENTICATION_METHOD":
-        case "ERR_POSTGRES_TLS_NOT_AVAILABLE":
-        case "ERR_POSTGRES_TLS_UPGRADE_FAILED":
-        case "ERR_POSTGRES_INVALID_SERVER_SIGNATURE":
-        case "ERR_POSTGRES_INVALID_SERVER_KEY":
-        case "ERR_POSTGRES_AUTHENTICATION_FAILED_PBKDF2":
-          // we can't retry these are authentication errors
-          return false;
-        default:
-          // we can retry
-          this.#doRetry();
-      }
-    }
-    return true;
   }
 }
 
 class PostgresAdapter
+  extends BaseSQLAdapter<
+    PooledPostgresConnection,
+    $ZigGeneratedClasses.PostgresSQLConnection,
+    $ZigGeneratedClasses.PostgresSQLQuery
+  >
   implements
     DatabaseAdapter<
       PooledPostgresConnection,
@@ -677,23 +448,8 @@ class PostgresAdapter
       $ZigGeneratedClasses.PostgresSQLQuery
     >
 {
-  public readonly connectionInfo: Bun.SQL.__internal.DefinedPostgresOrMySQLOptions;
-
-  public readonly connections: PooledPostgresConnection[];
-  public readonly readyConnections: Set<PooledPostgresConnection>;
-
-  public waitingQueue: Array<(err: Error | null, result: any) => void> = [];
-  public reservedQueue: Array<(err: Error | null, result: any) => void> = [];
-
-  public poolStarted: boolean = false;
-  public closed: boolean = false;
-  public totalQueries: number = 0;
-  public onAllQueriesFinished: (() => void) | null = null;
-
-  constructor(connectionInfo: Bun.SQL.__internal.DefinedPostgresOrMySQLOptions) {
-    this.connectionInfo = connectionInfo;
-    this.connections = new Array(connectionInfo.max);
-    this.readyConnections = new Set();
+  protected createPooledConnection(): PooledPostgresConnection {
+    return new PooledPostgresConnection(this.connectionInfo, this);
   }
 
   escapeIdentifier(str: string) {
@@ -720,26 +476,10 @@ class PostgresAdapter
       code: "ERR_POSTGRES_INVALID_TRANSACTION_STATE",
     });
   }
-  supportsReservedConnections() {
-    return true;
-  }
-
-  getConnectionForQuery(pooledConnection: PooledPostgresConnection) {
-    return pooledConnection.connection;
-  }
-
-  attachConnectionCloseHandler(connection: PooledPostgresConnection, handler: () => void): void {
-    // PostgreSQL pooled connections support onClose handlers
-    if (connection.onClose) {
-      connection.onClose(handler);
-    }
-  }
-
-  detachConnectionCloseHandler(connection: PooledPostgresConnection, handler: () => void): void {
-    // PostgreSQL pooled connections track queries
-    if (connection.queries) {
-      connection.queries.delete(handler);
-    }
+  unsafeTransactionError() {
+    return new PostgresError("Only use sql.begin, sql.reserved or max: 1", {
+      code: "ERR_POSTGRES_UNSAFE_TRANSACTION",
+    });
   }
 
   array(values: any[], typeNameOrID?: number | ArrayType): SQLArrayParameter {
@@ -779,28 +519,6 @@ class PostgresAdapter
     };
   }
 
-  validateTransactionOptions(options: string): { valid: boolean; error?: string } {
-    // The string is interpolated into `BEGIN ${options}`, so refuse anything that
-    // could terminate the statement or start a new one.
-    if (!/^[A-Za-z ,]*$/.test(options)) {
-      return {
-        valid: false,
-        error: "Transaction options can only contain letters, spaces, and commas.",
-      };
-    }
-    return { valid: true };
-  }
-
-  validateDistributedTransactionName(name: string): { valid: boolean; error?: string } {
-    if (name.indexOf("'") !== -1) {
-      return {
-        valid: false,
-        error: "Distributed transaction name cannot contain single quotes.",
-      };
-    }
-    return { valid: true };
-  }
-
   getCommitDistributedSQL(name: string): string {
     const validation = this.validateDistributedTransactionName(name);
     if (!validation.valid) {
@@ -818,16 +536,7 @@ class PostgresAdapter
   }
 
   createQueryHandle(sql: string, values: unknown[], flags: number) {
-    if (!(flags & SQLQueryFlags.allowUnsafeTransaction)) {
-      if (this.connectionInfo.max !== 1) {
-        const upperCaseSqlString = sql.toUpperCase().trim();
-        if (upperCaseSqlString.startsWith("BEGIN") || upperCaseSqlString.startsWith("START TRANSACTION")) {
-          throw new PostgresError("Only use sql.begin, sql.reserved or max: 1", {
-            code: "ERR_POSTGRES_UNSAFE_TRANSACTION",
-          });
-        }
-      }
-    }
+    this.checkUnsafeTransaction(sql, flags);
 
     return createPostgresQuery(
       sql,
@@ -839,567 +548,23 @@ class PostgresAdapter
     );
   }
 
-  maxDistribution() {
-    if (!this.waitingQueue.length) return 0;
-    const result = Math.ceil((this.waitingQueue.length + this.totalQueries) / this.connections.length);
-    return result ? result : 1;
+  getHelperCommand(query: string): SQLCommand {
+    return getHelperCommandFromDetect(query, false);
   }
 
-  flushConcurrentQueries() {
-    const maxDistribution = this.maxDistribution();
-    if (maxDistribution === 0) {
-      return;
-    }
-
-    while (true) {
-      const nonReservedConnections = Array.from(this.readyConnections || []).filter(
-        c => !(c.flags & PooledConnectionFlags.preReserved) && c.queryCount < maxDistribution,
-      );
-      if (nonReservedConnections.length === 0) {
-        return;
-      }
-      const orderedConnections = nonReservedConnections.sort((a, b) => a.queryCount - b.queryCount);
-      for (const connection of orderedConnections) {
-        const pending = this.waitingQueue.shift();
-        if (!pending) {
-          return;
-        }
-        connection.queryCount++;
-        this.totalQueries++;
-        pending(null, connection);
-      }
-    }
+  placeholder(index: number): string {
+    return "$" + index;
   }
 
-  release(connection: PooledPostgresConnection, connectingEvent: boolean = false) {
-    if (!connectingEvent) {
-      connection.queryCount--;
-      this.totalQueries--;
+  bindParam(value: unknown, binding_values: unknown[], index: number): string {
+    if (value instanceof SQLArrayParameter) {
+      binding_values.push(value.serializedValues);
+      return `$${index}::${value.arrayType}[] `;
     }
-    const currentQueryCount = connection.queryCount;
-    if (currentQueryCount == 0) {
-      connection.flags &= ~PooledConnectionFlags.reserved;
-      connection.flags &= ~PooledConnectionFlags.preReserved;
-    }
-    if (this.onAllQueriesFinished) {
-      // we are waiting for all queries to finish, lets check if we can call it
-      if (!this.hasPendingQueries()) {
-        this.onAllQueriesFinished();
-      }
-    }
-
-    if (connection.state !== PooledConnectionState.connected) {
-      // connection is not ready
-      if (connection.storedError) {
-        // this connection got a error but maybe we can wait for another
-
-        if (this.hasConnectionsAvailable()) {
-          return;
-        }
-
-        const waitingQueue = this.waitingQueue;
-        const reservedQueue = this.reservedQueue;
-
-        this.waitingQueue = [];
-        this.reservedQueue = [];
-        // we have no connections available so lets fails
-        for (const pending of waitingQueue) {
-          pending(connection.storedError, connection);
-        }
-        for (const pending of reservedQueue) {
-          pending(connection.storedError, connection);
-        }
-      }
-      return;
-    }
-
-    if (currentQueryCount == 0) {
-      // ok we can actually bind reserved queries to it
-      const pendingReserved = this.reservedQueue.shift();
-      if (pendingReserved) {
-        connection.flags |= PooledConnectionFlags.reserved;
-        connection.queryCount++;
-        this.totalQueries++;
-        // we have a connection waiting for a reserved connection lets prioritize it
-        pendingReserved(connection.storedError, connection);
-        return;
-      }
-    }
-    this.readyConnections.add(connection);
-    this.flushConcurrentQueries();
-  }
-
-  hasConnectionsAvailable() {
-    if (this.readyConnections?.size > 0) return true;
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection && connection.state !== PooledConnectionState.closed) {
-          // some connection is connecting or connected
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  hasPendingQueries() {
-    if (this.waitingQueue.length > 0 || this.reservedQueue.length > 0) return true;
-    if (this.poolStarted) {
-      return this.totalQueries > 0;
-    }
-    return false;
-  }
-  isConnected() {
-    if (this.readyConnections?.size > 0) {
-      return true;
-    }
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection.state === PooledConnectionState.connected) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-  flush() {
-    if (this.closed) {
-      return;
-    }
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection.state === PooledConnectionState.connected) {
-          connection.connection?.flush();
-        }
-      }
-    }
-  }
-
-  async #close() {
-    let pending;
-    while ((pending = this.waitingQueue.shift())) {
-      pending(this.connectionClosedError(), null);
-    }
-    while (this.reservedQueue.length > 0) {
-      const pendingReserved = this.reservedQueue.shift();
-      if (pendingReserved) {
-        pendingReserved(this.connectionClosedError(), null);
-      }
-    }
-
-    const promises: Array<Promise<any>> = [];
-
-    if (this.poolStarted) {
-      this.poolStarted = false;
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        switch (connection.state) {
-          case PooledConnectionState.pending:
-            {
-              const { promise, resolve } = Promise.withResolvers();
-              connection.onFinish = resolve;
-              promises.push(promise);
-              connection.connection?.close();
-            }
-            break;
-
-          case PooledConnectionState.connected:
-            {
-              const { promise, resolve } = Promise.withResolvers();
-              connection.onFinish = resolve;
-              promises.push(promise);
-              connection.connection?.close();
-            }
-            break;
-        }
-        // clean connection reference
-        // @ts-ignore
-        this.connections[i] = null;
-      }
-    }
-
-    this.readyConnections.clear();
-    this.waitingQueue.length = 0;
-    return Promise.all(promises);
-  }
-
-  async close(options?: { timeout?: number }): Promise<void> {
-    if (this.closed) {
-      return;
-    }
-
-    let timeout = options?.timeout;
-    if (timeout) {
-      timeout = Number(timeout);
-      if (timeout > 2 ** 31 || timeout < 0 || timeout !== timeout) {
-        throw $ERR_INVALID_ARG_VALUE("options.timeout", timeout, "must be a non-negative integer less than 2^31");
-      }
-
-      this.closed = true;
-      if (timeout === 0 || !this.hasPendingQueries()) {
-        // close immediately
-        await this.#close();
-        return;
-      }
-
-      const { promise, resolve } = Promise.withResolvers<void>();
-      const timer = setTimeout(() => {
-        // timeout is reached, lets close and probably fail some queries
-        this.#close().finally(resolve);
-      }, timeout * 1000);
-      timer.unref(); // dont block the event loop
-
-      this.onAllQueriesFinished = () => {
-        clearTimeout(timer);
-        // everything is closed, lets close the pool
-        this.#close().finally(resolve);
-      };
-
-      return promise;
-    } else {
-      this.closed = true;
-      if (!this.hasPendingQueries()) {
-        // close immediately
-        await this.#close();
-        return;
-      }
-
-      // gracefully close the pool
-      const { promise, resolve } = Promise.withResolvers<void>();
-
-      this.onAllQueriesFinished = () => {
-        // everything is closed, lets close the pool
-        this.#close().finally(resolve);
-      };
-
-      return promise;
-    }
-  }
-
-  /**
-   * @param {function} onConnected - The callback function to be called when the connection is established.
-   * @param {boolean} reserved - Whether the connection is reserved, if is reserved the connection will not be released until release is called, if not release will only decrement the queryCount counter
-   */
-  connect(onConnected: (err: Error | null, result: any) => void, reserved: boolean = false) {
-    if (this.closed) {
-      return onConnected(this.connectionClosedError(), null);
-    }
-
-    if (!this.readyConnections || this.readyConnections.size === 0) {
-      // no connection ready lets make some
-      let retry_in_progress = false;
-      let all_closed = true;
-      let storedError: Error | null = null;
-
-      if (this.poolStarted) {
-        // we already started the pool
-        // lets check if some connection is available to retry
-        const pollSize = this.connections.length;
-        for (let i = 0; i < pollSize; i++) {
-          const connection = this.connections[i];
-          // we need a new connection and we have some connections that can retry
-          if (connection.state === PooledConnectionState.closed) {
-            if (connection.retry()) {
-              // lets wait for connection to be released
-              if (!retry_in_progress) {
-                // avoid adding to the queue twice, we wanna to retry every available pool connection
-                retry_in_progress = true;
-                if (reserved) {
-                  // we are not sure what connection will be available so we dont pre reserve
-                  this.reservedQueue.push(onConnected);
-                } else {
-                  this.waitingQueue.push(onConnected);
-                }
-              }
-            } else {
-              // we have some error, lets grab it and fail if unable to start a connection
-              storedError = connection.storedError;
-            }
-          } else {
-            // we have some pending or open connections
-            all_closed = false;
-          }
-        }
-        if (!all_closed && !retry_in_progress) {
-          // is possible to connect because we have some working connections, or we are just without network for some reason
-          // wait for connection to be released or fail
-          if (reserved) {
-            // we are not sure what connection will be available so we dont pre reserve
-            this.reservedQueue.push(onConnected);
-          } else {
-            this.waitingQueue.push(onConnected);
-          }
-        } else if (!retry_in_progress) {
-          // impossible to connect or retry
-          onConnected(storedError ?? this.connectionClosedError(), null);
-        }
-        return;
-      }
-      // we never started the pool, lets start it
-      if (reserved) {
-        this.reservedQueue.push(onConnected);
-      } else {
-        this.waitingQueue.push(onConnected);
-      }
-      this.poolStarted = true;
-      const pollSize = this.connections.length;
-      // pool is always at least 1 connection
-      const firstConnection = new PooledPostgresConnection(this.connectionInfo, this);
-      this.connections[0] = firstConnection;
-      if (reserved) {
-        firstConnection.flags |= PooledConnectionFlags.preReserved; // lets pre reserve the first connection
-      }
-      for (let i = 1; i < pollSize; i++) {
-        this.connections[i] = new PooledPostgresConnection(this.connectionInfo, this);
-      }
-      return;
-    }
-    if (reserved) {
-      let connectionWithLeastQueries: PooledPostgresConnection | null = null;
-      let leastQueries = Infinity;
-      for (const connection of this.readyConnections || []) {
-        if (connection.flags & PooledConnectionFlags.preReserved || connection.flags & PooledConnectionFlags.reserved)
-          continue;
-        const queryCount = connection.queryCount;
-        if (queryCount > 0) {
-          if (queryCount < leastQueries) {
-            leastQueries = queryCount;
-            connectionWithLeastQueries = connection;
-          }
-          continue;
-        }
-        connection.flags |= PooledConnectionFlags.reserved;
-        connection.queryCount++;
-        this.totalQueries++;
-        this.readyConnections?.delete(connection);
-        onConnected(null, connection);
-        return;
-      }
-
-      if (connectionWithLeastQueries) {
-        // lets mark the connection with the least queries as preReserved if any
-        connectionWithLeastQueries.flags |= PooledConnectionFlags.preReserved;
-      }
-
-      // no connection available to be reserved lets wait for a connection to be released
-      this.reservedQueue.push(onConnected);
-    } else {
-      this.waitingQueue.push(onConnected);
-      this.flushConcurrentQueries();
-    }
-  }
-
-  normalizeQuery(strings: string | TemplateStringsArray, values: unknown[], binding_idx = 1): [string, unknown[]] {
-    // This function handles array values in single fields:
-    // - JSON/JSONB are the only field types that can be arrays themselves, so we serialize them
-    // - SQL array field types (e.g., INTEGER[], TEXT[]) require the sql.array() helper
-    // - All other types are handled natively
-
-    if (typeof strings === "string") {
-      // identifier or unsafe query
-      return [strings, values || []];
-    }
-
-    if (!$isArray(strings)) {
-      // we should not hit this path
-      throw new SyntaxError("Invalid query: SQL Fragment cannot be executed or was misused");
-    }
-
-    const str_len = strings.length;
-    if (str_len === 0) {
-      return ["", []];
-    }
-
-    let binding_values: any[] = [];
-    let query = "";
-
-    for (let i = 0; i < str_len; i++) {
-      const string = strings[i];
-
-      if (typeof string === "string") {
-        query += string;
-
-        if (values.length > i) {
-          const value = values[i];
-
-          if (value instanceof Query) {
-            const q = value as Query<any, any>;
-            const [sub_query, sub_values] = this.normalizeQuery(q[_strings], q[_values], binding_idx);
-
-            query += sub_query;
-            for (let j = 0; j < sub_values.length; j++) {
-              binding_values.push(sub_values[j]);
-            }
-            binding_idx += sub_values.length;
-          } else if (value instanceof SQLHelper) {
-            const command = detectCommand(query);
-            // only selectIn, insert, update, updateSet are allowed
-            if (command === SQLCommand.none || command === SQLCommand.where) {
-              throw new SyntaxError("Helpers are only allowed for INSERT, UPDATE and IN commands");
-            }
-            const { columns, value: items } = value as SQLHelper;
-            const columnCount = columns.length;
-            if (columnCount === 0 && command !== SQLCommand.in) {
-              throw new SyntaxError(`Cannot ${commandToString(command)} with no columns`);
-            }
-            const lastColumnIndex = columns.length - 1;
-
-            if (command === SQLCommand.insert) {
-              //
-              // insert into users ${sql(users)} or insert into users ${sql(user)}
-              //
-
-              // Build column list while determining which columns have at least one defined value
-              const { definedColumns, columnsSql } = buildDefinedColumnsAndQuery(
-                columns,
-                items,
-                this.escapeIdentifier.bind(this),
-              );
-
-              const definedColumnCount = definedColumns.length;
-              if (definedColumnCount === 0) {
-                throw new SyntaxError("Insert needs to have at least one column with a defined value");
-              }
-              const lastDefinedColumnIndex = definedColumnCount - 1;
-
-              query += columnsSql;
-              if ($isArray(items)) {
-                const itemsCount = items.length;
-                const lastItemIndex = itemsCount - 1;
-                for (let j = 0; j < itemsCount; j++) {
-                  query += "(";
-                  const item = items[j];
-                  for (let k = 0; k < definedColumnCount; k++) {
-                    const column = definedColumns[k];
-                    const columnValue = item[column];
-                    query += `$${binding_idx++}${k < lastDefinedColumnIndex ? ", " : ""}`;
-                    // If this item has undefined for a column that other items defined, use null
-                    binding_values.push(typeof columnValue === "undefined" ? null : columnValue);
-                  }
-                  if (j < lastItemIndex) {
-                    query += "),";
-                  } else {
-                    query += ") "; // the user can add RETURNING * or RETURNING id
-                  }
-                }
-              } else {
-                query += "(";
-                const item = items;
-                for (let j = 0; j < definedColumnCount; j++) {
-                  const column = definedColumns[j];
-                  const columnValue = item[column];
-                  query += `$${binding_idx++}${j < lastDefinedColumnIndex ? ", " : ""}`;
-                  binding_values.push(columnValue);
-                }
-                query += ") "; // the user can add RETURNING * or RETURNING id
-              }
-            } else if (command === SQLCommand.in) {
-              // SELECT * FROM users WHERE id IN (${sql([1, 2, 3])})
-              if (!$isArray(items)) {
-                throw new SyntaxError("An array of values is required for WHERE IN helper");
-              }
-              const itemsCount = items.length;
-              const lastItemIndex = itemsCount - 1;
-              query += "(";
-              for (let j = 0; j < itemsCount; j++) {
-                query += `$${binding_idx++}${j < lastItemIndex ? ", " : ""}`;
-                if (columnCount > 0) {
-                  // we must use a key from a object
-                  if (columnCount > 1) {
-                    // we should not pass multiple columns here
-                    throw new SyntaxError("Cannot use WHERE IN helper with multiple columns");
-                  }
-                  // SELECT * FROM users WHERE id IN (${sql(users, "id")})
-                  const value = items[j];
-                  if (typeof value === "undefined") {
-                    binding_values.push(null);
-                  } else {
-                    const value_from_key = value[columns[0]];
-
-                    if (typeof value_from_key === "undefined") {
-                      binding_values.push(null);
-                    } else {
-                      binding_values.push(value_from_key);
-                    }
-                  }
-                } else {
-                  const value = items[j];
-                  if (typeof value === "undefined") {
-                    binding_values.push(null);
-                  } else {
-                    binding_values.push(value);
-                  }
-                }
-              }
-              query += ") "; // more conditions can be added after this
-            } else {
-              // UPDATE users SET ${sql({ name: "John", age: 31 })} WHERE id = 1
-              let item;
-              if ($isArray(items)) {
-                if (items.length > 1) {
-                  throw new SyntaxError("Cannot use array of objects for UPDATE");
-                }
-                item = items[0];
-              } else {
-                item = items;
-              }
-              // no need to include if is updateSet
-              if (command === SQLCommand.update) {
-                query += " SET ";
-              }
-              let hasValues = false;
-              for (let i = 0; i < columnCount; i++) {
-                const column = columns[i];
-                const columnValue = item[column];
-                if (typeof columnValue === "undefined") {
-                  // skip undefined values, this is the expected behavior in JS
-                  continue;
-                }
-                hasValues = true;
-                query += `${this.escapeIdentifier(column)} = $${binding_idx++}${i < lastColumnIndex ? ", " : ""}`;
-                binding_values.push(columnValue);
-              }
-              if (query.endsWith(", ")) {
-                // we got an undefined value at the end, lets remove the last comma
-                query = query.substring(0, query.length - 2);
-              }
-              if (!hasValues) {
-                throw new SyntaxError("Update needs to have at least one column");
-              }
-              // the user can add where clause after this
-              query += " ";
-            }
-          } else if (value instanceof SQLArrayParameter) {
-            query += `$${binding_idx++}::${value.arrayType}[] `;
-            binding_values.push(value.serializedValues);
-          } else {
-            query += `$${binding_idx++} `;
-            if (typeof value === "undefined") {
-              binding_values.push(null);
-            } else {
-              binding_values.push(value);
-            }
-          }
-        }
-      } else {
-        throw new SyntaxError("Invalid query: SQL Fragment cannot be executed or was misused");
-      }
-    }
-
-    return [query, binding_values];
+    return pushBindParam(this, value, binding_values, index);
   }
 }
 
 export default {
   PostgresAdapter,
-  SQLCommand,
-  commandToString,
-  detectCommand,
 };
