+++ src/js/internal/sql/sqlite.ts
@@ -1,13 +1,16 @@
 import type * as BunSQLiteModule from "bun:sqlite";
 import type { BaseQueryHandle, Query, SQLQueryResultMode } from "./query";
-import type { ArrayType, DatabaseAdapter, OnConnected, SQLArrayParameter, SQLHelper, SQLResultArray } from "./shared";
+import type {
+  ArrayType,
+  DatabaseAdapter,
+  OnConnected,
+  SQLArrayParameter,
+  SQLCommand as SharedSQLCommand,
+  SQLResultArray,
+} from "./shared";
 
-const { SQLHelper, SQLResultArray, buildDefinedColumnsAndQuery } = require("internal/sql/shared");
-const {
-  Query,
-  SQLQueryResultMode,
-  symbols: { _strings, _values },
-} = require("internal/sql/query");
+const { SQLResultArray, normalizeQuery, pushBindParam } = require("internal/sql/shared");
+const { SQLQueryResultMode } = require("internal/sql/query");
 const { SQLiteError } = require("internal/sql/errors");
 
 let lazySQLiteModule: typeof BunSQLiteModule;
@@ -377,204 +380,38 @@ class SQLiteAdapter implements DatabaseAdapter<BunSQLiteModule.Database, BunSQLi
     });
   }
   normalizeQuery(strings: string | TemplateStringsArray, values: unknown[], binding_idx = 1): [string, unknown[]] {
-    if (typeof strings === "string") {
-      // identifier or unsafe query
-      return [strings, values || []];
+    return normalizeQuery(this, strings, values, binding_idx);
+  }
+
+  // SQLite uses ? for placeholders, not $1, $2, etc.
+  placeholder(_index: number): string {
+    return "?";
+  }
+
+  bindParam(value: unknown, binding_values: unknown[], index: number): string {
+    return pushBindParam(this, value, binding_values, index);
+  }
+
+  getHelperCommand(query: string): SharedSQLCommand {
+    // when partial is true we stop on the first command we find
+    const { command } = parseSQLQuery(query, true);
+
+    // only selectIn, insert, update, updateSet are allowed
+    if (command === SQLCommand.none || command === SQLCommand.where) {
+      throw new SyntaxError("Helpers are only allowed for INSERT, UPDATE and WHERE IN commands");
     }
+    // the local SQLCommand enum is numerically identical to the shared one
+    return command as unknown as SharedSQLCommand;
+  }
 
-    if (!$isArray(strings)) {
-      // we should not hit this path
-      throw new SyntaxError("Invalid query: SQL Fragment cannot be executed or was misused");
+  isUpsertUpdate(_query: string): boolean {
+    return false;
+  }
+
+  throwIfUpdateEmpty(query: string, _hasValues: boolean): void {
+    if (query.endsWith("SET ")) {
+      throw new SyntaxError("Update needs to have at least one column");
     }
-
-    const str_len = strings.length;
-    if (str_len === 0) {
-      return ["", []];
-    }
-
-    let binding_values: any[] = [];
-    let query = "";
-
-    for (let i = 0; i < str_len; i++) {
-      const string = strings[i];
-
-      if (typeof string === "string") {
-        query += string;
-
-        if (values.length > i) {
-          const value = values[i];
-
-          if (value instanceof Query) {
-            const q = value as Query<any, any>;
-            const [sub_query, sub_values] = this.normalizeQuery(q[_strings], q[_values], binding_idx);
-
-            query += sub_query;
-            for (let j = 0; j < sub_values.length; j++) {
-              binding_values.push(sub_values[j]);
-            }
-            binding_idx += sub_values.length;
-          } else if (value instanceof SQLHelper) {
-            // when partial is true we stop on the first command we find
-            const { command } = parseSQLQuery(query, true);
-
-            // only selectIn, insert, update, updateSet are allowed
-            if (command === SQLCommand.none || command === SQLCommand.where) {
-              throw new SyntaxError("Helpers are only allowed for INSERT, UPDATE and WHERE IN commands");
-            }
-            const { columns, value: items } = value as SQLHelper;
-            const columnCount = columns.length;
-            if (columnCount === 0 && command !== SQLCommand.in) {
-              throw new SyntaxError(`Cannot ${commandToString(command)} with no columns`);
-            }
-            const lastColumnIndex = columns.length - 1;
-
-            if (command === SQLCommand.insert) {
-              //
-              // insert into users ${sql(users)} or insert into users ${sql(user)}
-              //
-
-              // Build column list while determining which columns have at least one defined value
-              const { definedColumns, columnsSql } = buildDefinedColumnsAndQuery(
-                columns,
-                items,
-                this.escapeIdentifier.bind(this),
-              );
-
-              const definedColumnCount = definedColumns.length;
-              if (definedColumnCount === 0) {
-                throw new SyntaxError("Insert needs to have at least one column with a defined value");
-              }
-              const lastDefinedColumnIndex = definedColumnCount - 1;
-
-              query += columnsSql;
-              if ($isArray(items)) {
-                const itemsCount = items.length;
-                const lastItemIndex = itemsCount - 1;
-                for (let j = 0; j < itemsCount; j++) {
-                  query += "(";
-                  const item = items[j];
-                  for (let k = 0; k < definedColumnCount; k++) {
-                    const column = definedColumns[k];
-                    const columnValue = item[column];
-                    // SQLite uses ? for placeholders, not $1, $2, etc.
-                    query += `?${k < lastDefinedColumnIndex ? ", " : ""}`;
-                    // If this item has undefined for a column that other items defined, use null
-                    binding_values.push(typeof columnValue === "undefined" ? null : columnValue);
-                  }
-                  if (j < lastItemIndex) {
-                    query += "),";
-                  } else {
-                    query += ") "; // the user can add RETURNING * or RETURNING id
-                  }
-                }
-              } else {
-                query += "(";
-                const item = items;
-                for (let j = 0; j < definedColumnCount; j++) {
-                  const column = definedColumns[j];
-                  const columnValue = item[column];
-                  // SQLite uses ? for placeholders
-                  query += `?${j < lastDefinedColumnIndex ? ", " : ""}`;
-                  binding_values.push(columnValue);
-                }
-                query += ") "; // the user can add RETURNING * or RETURNING id
-              }
-            } else if (command === SQLCommand.in) {
-              // SELECT * FROM users WHERE id IN (${sql([1, 2, 3])})
-              if (!$isArray(items)) {
-                throw new SyntaxError("An array of values is required for WHERE IN helper");
-              }
-              const itemsCount = items.length;
-              const lastItemIndex = itemsCount - 1;
-              query += "(";
-              for (let j = 0; j < itemsCount; j++) {
-                // SQLite uses ? for placeholders
-                query += `?${j < lastItemIndex ? ", " : ""}`;
-                if (columnCount > 0) {
-                  // we must use a key from a object
-                  if (columnCount > 1) {
-                    // we should not pass multiple columns here
-                    throw new SyntaxError("Cannot use WHERE IN helper with multiple columns");
-                  }
-                  // SELECT * FROM users WHERE id IN (${sql(users, "id")})
-                  const value = items[j];
-                  if (typeof value === "undefined") {
-                    binding_values.push(null);
-                  } else {
-                    const value_from_key = value[columns[0]];
-
-                    if (typeof value_from_key === "undefined") {
-                      binding_values.push(null);
-                    } else {
-                      binding_values.push(value_from_key);
-                    }
-                  }
-                } else {
-                  const value = items[j];
-                  if (typeof value === "undefined") {
-                    binding_values.push(null);
-                  } else {
-                    binding_values.push(value);
-                  }
-                }
-              }
-              query += ") "; // more conditions can be added after this
-            } else {
-              // UPDATE users SET ${sql({ name: "John", age: 31 })} WHERE id = 1
-              let item;
-              if ($isArray(items)) {
-                if (items.length > 1) {
-                  throw new SyntaxError("Cannot use array of objects for UPDATE");
-                }
-                item = items[0];
-              } else {
-                item = items;
-              }
-              // no need to include if is updateSet
-              if (command === SQLCommand.update) {
-                query += " SET ";
-              }
-              for (let i = 0; i < columnCount; i++) {
-                const column = columns[i];
-                const columnValue = item[column];
-                if (typeof columnValue === "undefined") {
-                  // skip undefined values, this is the expected behavior in JS
-                  continue;
-                }
-                // SQLite uses ? for placeholders
-                query += `${this.escapeIdentifier(column)} = ?${i < lastColumnIndex ? ", " : ""}`;
-                if (typeof columnValue === "undefined") {
-                  binding_values.push(null);
-                } else {
-                  binding_values.push(columnValue);
-                }
-              }
-              if (query.endsWith(", ")) {
-                // we got an undefined value at the end, lets remove the last comma
-                query = query.substring(0, query.length - 2);
-              }
-              if (query.endsWith("SET ")) {
-                throw new SyntaxError("Update needs to have at least one column");
-              }
-              // the user can add where clause after this
-              query += " ";
-            }
-          } else {
-            // SQLite uses ? for placeholders
-            query += `? `;
-            if (typeof value === "undefined") {
-              binding_values.push(null);
-            } else {
-              binding_values.push(value);
-            }
-          }
-        }
-      } else {
-        throw new SyntaxError("Invalid query: SQL Fragment cannot be executed or was misused");
-      }
-    }
-
-    return [query, binding_values];
   }
 
   connect(onConnected: OnConnected<BunSQLiteModule.Database>, reserved?: boolean) {
